requirements.txt
aiosmtplib==1.1.7
anyio==3.6.1
APScheduler==3.9.1.post1
attrs==22.1.0
Authlib==1.1.0
base58==2.1.1
beanie==1.12.0
blinker==1.5
boto3==1.26.37
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
crypto==1.4.1
cryptography==38.0.3
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
fastapi==0.85.0
fastapi-mail==1.1.5
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
googleapis-common-protos==1.56.4
h11==0.12.0
httpcore==0.15.0
httplib2==0.21.0
httpx==0.23.0
idna==3.4
iniconfig==1.1.1
Jinja2==3.1.2
jose==1.0.0
loguru==0.6.0
MarkupSafe==2.1.1
motor==3.0.0
multidict==6.0.2
Naked==0.1.32
nanoid==2.0.0
oauthlib==3.2.2
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
pony==0.7.16
protobuf==4.21.9
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.10.2
pymongo==4.2.0
pyparsing==3.0.9
pytest==7.1.3
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
random-object-id==2.0.0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
shellescape==3.8.1
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
toml==0.10.2
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.18.3
yarl==1.8.1

boto3~=1.26.37
botocore~=1.29.37
elastic-apm